Move Remote Feed Item to it's own file
