+++ EssentialDeveloper/EssentialDeveloper/FeedAPI/FeedItemsMapper.swift
@@ -1,12 +1,5 @@
 import Foundation
 
-internal struct RemoteFeedItem: Decodable {
-    internal let id: UUID
-    internal let description: String?
-    internal let location: String?
-    internal let image: URL
-}
-
 internal final class FeedItemsMapper {
     private struct Root: Decodable {
         let items: [RemoteFeedItem]
