+++ EssentialDeveloper/EssentialDeveloper/FeedAPI/RemoteFeedItem.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+internal struct RemoteFeedItem: Decodable {
+    internal let id: UUID
+    internal let description: String?
+    internal let location: String?
+    internal let image: URL
+}
